Fixes #432 - Create association on locale change
According to the docs, calling MyPage.instance.header
should get the header text for current locale.
However, the associations are set up in a singleton
for a default locale and never modified.

The language-specific methods like MyPage.instance.header_en
still work, but it would be nice to have fae handle the locale
for us.

This change checks for languages in fae_fields for page
and dynamically changes the association if the locale change
is detected.